+++ backend/src/main/java/com/devdive/backend/post/application/service/PostService.java
@@ -3,7 +3,7 @@ package com.devdive.backend.post.application.service;
 import com.devdive.backend.post.application.dto.PostCreateRequestDto;
 import com.devdive.backend.post.application.dto.PostViewDto;
 import com.devdive.backend.post.application.port.in.PostUseCase;
-import com.devdive.backend.post.application.port.out.persistence.LoadPostPort;
+import com.devdive.backend.post.application.port.out.LoadPostPort;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
